refactor(test): group remaining automation coverage
