+++ test/automation/e2e/no-unit-blocks-in-live-e2e.test.ts
@@ -6,7 +6,7 @@ import { describe, expect, it } from "vitest";
 import {
   findLiveUnitBlocks,
   formatViolations,
-} from "../scripts/checks/no-unit-blocks-in-live-e2e.mts";
+} from "../../../scripts/checks/no-unit-blocks-in-live-e2e.mts";
 
 const FILE = "test/e2e/live/example.test.ts";
