+++ test/automation/releases/changelog-docs.test.ts
@@ -6,9 +6,9 @@ import path from "node:path";
 import { describe, expect, it } from "vitest";
 import { parse } from "yaml";
 
-import { extractMarkdownLinks } from "../scripts/check-docs-published-routes.mts";
+import { extractMarkdownLinks } from "../../../scripts/check-docs-published-routes.mts";
 
-const docsDir = path.join(import.meta.dirname, "..", "docs");
+const docsDir = path.join(import.meta.dirname, "../../..", "docs");
 const changelogDir = path.join(docsDir, "changelog");
 const mdxSpdxHeader = `{/*
  * SPDX-FileCopyrightText: Copyright (c) 2026 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
@@ -193,7 +193,7 @@ describe("Fern changelog documentation", () => {
     expect(overview).not.toContain("Component Version Policy");
     expect(updateSandboxes).toContain("## Understand Agent Version Pins");
     expect(commands).toContain(
-      "../manage-sandboxes/operate-sandboxes/update-sandboxes#understand-agent-version-pins",
+      "../../../manage-sandboxes/operate-sandboxes/update-sandboxes#understand-agent-version-pins",
     );
   });
 
@@ -223,7 +223,7 @@ describe("Fern changelog documentation", () => {
     variants.forEach((variant) => {
       expect(variant.layout?.filter((node) => node.changelog)).toEqual([
         {
-          changelog: "./changelog",
+          changelog: "../.././changelog",
           title: "Release Notes",
           slug: "release-notes",
         },
